GraphSEIR.jl Documentation

Disease simulation package for network diffusion SEIR model of disease transmission. In this package we use terminology networks and graphs, nodes and vertices interchangeably.

Graph Functions

GraphSEIR.load_graph — Method

load_graph(vertices_filename::String, edges_filename::String)

Return a MetaGraph corresponding to attributes in the corresponding vertices_filename and edges_filename.

GraphSEIR.load_edgefile — Function

load_edgefile(vlist::Array{Int}, edges_filename::String; scale::Float64=1.0)

Returns an edge list and weight matrix for the attributes in the corresponding edges_filename with weights scaled by option scale.

GraphSEIR.swap_edges! — Function

swap_edges!(g::SEIRGraph, edge_list::Dict, weight_matrix::SparseMatrixCSC)

Replaces edges of graph g inplace with edges from edge_list and weights from weight_matrix.

GraphSEIR.vertex_weight — Function

vertex_weight(g, v, filter_household=false)

Computes weighted in-degree of a vertex v in graph g. Set filter_household flag to true to filter household edges from computation.

Disease Functions

GraphSEIR.count — Method

count(  g::SEIRGraph, s::SEIRStatus)

Returns the number of vertices with state s.

GraphSEIR.count — Method

count(  g::SEIRGraph, c::Int, s::SEIRStatus)

Returns the number of vertices in community c that have state s.

GraphSEIR.count — Method

count(  g::SEIRGraph)

Returns the total number of vertices in the graph.

GraphSEIR.counts — Method

counts(  g::SEIRGraph)

Returns the total number of vertices in each state S,E,I,R.

GraphSEIR.counts_by_community — Method

counts_by_community(g::SEIRGraph, s::SEIRStatus)

Returns a list of the number of vertices in each community that have state s.

GraphSEIR.get_indices — Method

get_indices(  g::SEIRGraph, s::SEIRStatus)

Returns an iterator of the vertices in state s.

GraphSEIR.get_indices — Method

get_indices(  g::SEIRGraph, c::Int, s::SEIRStatus)

Returns an iterator of all vertices in community c that have state s.

GraphSEIR.get_indices — Method

get_indices(  g::SEIRGraph)

Returns iterators of the vertices in each state S,E,I,R.

GraphSEIR.SEIRTransitionHistory — Type

SEIRTransitionHistory(SEE::T, SEI::T, EI::T, IR::T, S::T, E::T, I::T, R::T) where T <: Int

Structure for tracking SEIR transition history.

GraphSEIR.CompartmentalTracker — Method

CompartmentalTracker(communities::Array{Int})

Create dictionary for tracking compartmental and SEIR statistics for input communities.

GraphSEIR.SEIRInfectionTree — Type

SEIRInfectionTree

Tree detailing the SEIR infection history.

GraphSEIR.initialize_with! — Function

initialize_with!(g::SEIRGraph, p::Function, rng::AbstractRNG)

Initialize statuses of vertices in graph g using density function p(g,v) that defines the probability of a vertex v being in each SEIR state. Updates are stored in infection tree t.

GraphSEIR.expose! — Function

expose!(g::SEIRGraph, p::AbstractFloat, rng::AbstractRNG)

Randomly initialize vertices to statuse Exposed(E) with probability p.

GraphSEIR.reset! — Function

reset!(g::SEIRGraph, rng::AbstractRNG)

Set the status of all vertices to Susceptible(S).

GraphSEIR.expose_communities! — Function

expose_communities!(g::SEIRGraph, p::AbstractFloat, nc::Int, rng::AbstractRNG)

Expose p fraction of individuals in the entire graph, localized to nc communities.

GraphSEIR.expose_communities — Function

expose_communities(g::SEIRGraph, v::Int, p::AbstractFloat, communities::Vector{Int})

Graph-vertex function for initializing graph with exposed communities.

GraphSEIR.add_new_infection! — Function

add_new_infection!(t::SEIRInfectionTree, time_index::Int64, src::Dict, dst::Dict)

Add a new infection to the infection tree t given the src and dst vertices at time_index.

GraphSEIR.transition! — Function

transition!(;g::SEIRGraph, params::Union{Nothing, SEIRParameters} = nothing, rng::AbstractRNG, tracker::Union{Nothing, Dict} = nothing, t::Union{Nothing, SEIRInfectionTree} = nothing, time_index::Int = 0)

Transition graph g one time step based on the input SEIR parameters and tracking options. GlobalParameters and CommunityParameters types use the fast vectorized transition.

GraphSEIR.full_transition! — Function

full_transition!(g::SEIRGraph, params::SEIRParameters, rng::AbstractRNG, tracker::Union{Nothing,Dict}, t::SEIRInfectionTree, time_index::Int)

Transition the SEIRGraph g one time step and fully track infection tree.

GraphSEIR.short_circuit_transition! — Function

short_circuit_transition!(g::SEIRGraph, params::CommunityPairParameters, rng::AbstractRNG, tracker::Union{Nothing,Dict} = nothing)

Transition the SEIRGraph g one time step using CommunityPairParameters with short circuit evaluation. Cannot track an infection tree.

GraphSEIR.vectorized_transition! — Function

vectorized_transition!(g::SEIRGraph, rng::AbstractRNG)

Transition the SEIRGraph g using GlobalParameters or CommunityParameters. Cannot track infection type or infection tree.

GraphSEIR.mdp_transition! — Function

mdp_transition!(;g::SEIRGraph, rng::AbstractRNG)

Specialized transition function for use in EpidemicGraphMDPs planning.

Community Ordering Heuristics

GraphSEIR.transition_likelihood — Function

transition_likelihood(g::SEIRGraph)

Return the likelihood of all nodes transitioning to the following SEIR status given the node's current SEIR status, neighbors, and SEIR parameters.

GraphSEIR.community_expected_exposure — Function

community_expected_exposure(g::GraphSEIR, c::Int)

Return the expected number of individuals in community c that will become exposed at the next time step. (complex binomial - MLE)

GraphSEIR.community_expected_infection — Function

community_expected_infection(g::GraphSEIR, c::Int)

Return the expected number of individuals in community c that will become infected at the next time step. (binomial - MLE)

GraphSEIR.community_expected_recovery — Function

community_expected_recovery(g::GraphSEIR, c::Int)

Return the expectation of the number of individuals in community c that will become recovered at the next time step. (binomial - MLE)

GraphSEIR.community_expected_exposures — Function

community_expected_exposures(g::GraphSEIR)

Return the expected number of individuals in each community that will become exposed at the next time step. (complex binomial - MLE) Same function naming convention exists for infections and recoveries.

GraphSEIR.exposure_subgraph — Function

exposure_subgraph(g::SEIRGraph)

Return the subgraph corresponding to S-E & S-I edges.

GraphSEIR.epidemic_mobility_weight_matrix — Function

epidemic_mobility_weight_matrix(g::SEIRGraph)

Get epidemic mobility weight matrix for corresponding SEIRGraph g.

GraphSEIR.eigenvectors — Function

eigenvectors(g::SEIRGraph, n::Int)

Return the n leading left (u) and right (v) eigenvectors of the epidemic-mobility weight matrix of the SEIRGraph g.

GraphSEIR.eigenscore — Function

eigenscore(e::AbstractEdge, u::Vector, v::Vector)

Return the eigenscore u(i)*v(j) for edge e: i -> j, with left u and right v eigenvectors of the epidemic-mobility weight matrix.

GraphSEIR.community_eigenscores — Function

community_eigenscores(g::SEIRGraph)

Return the sum of eigenscores for all non-household edges in each community.

SEIR Parameter Functions

GraphSEIR.SEIRParameters — Type

SEIRParameters{T}

Abstract type for SEIR parameters.

GraphSEIR.GlobalParameters — Type

GlobalParameters(βE::T, βI::T, γ::T, λ::T)

Structure for global scalar SEIR parameters.

GraphSEIR.CommunityParameters — Type

CommunityParams(βE::Vector{T} βI::Vector{T}, γ::T, λ::T, c::Int) where T <: Real

Structure for individual community-based SEIRParameters constructor given vector β values and c communities.

Can be constructed with scalar β values to set all communities to the same β values.

GraphSEIR.CommunityPairParameters — Type

CommunityPairParameters(βE::Matrix{T}, βI::Matrix{T}, γ::T, λ::T)

Structure for community-pair based SEIR parameters.

Can be constructed with scalar β values to set all community-pair β values.

GraphSEIR.R_i — Function

R_i(g, β_e, vs = []; γ = 0.2)

Compute expected number of infections generated by each vertex for specified β and γ values.

GraphSEIR.compute_R0 — Function

compute_R0(g, β_e; α=1.0, γ=0.2, filter_household = false, vweights = nothing)

Approximate the reproduction rate of a disease for a given graph g, infection parameters β and γ, and top proportion of nodes by degree α. vweights can be supplied if pre-computed.

GraphSEIR.get_betas — Function

get_betas(g, R0; α=1.0, γ=0.2, λ=0.2, p = 1/5, filter_household = false, range = [-5,0])

Fit β_e and β_i parameters for a graph g with disease reproduction rate R0. range is the upper and lower bound as an order of magnitude for fitting.

Analysis Helper Functions

GraphSEIR.format_data — Function

format_data(tracker, g, pv, params, save_dir::Union{Nothing, String} = nothing)

Format simulation output from tracker into global SEIR statistics and community SEIR statistics. Saves as JSON files to save_dir if provided.

Edge Filter Helper Functions

GraphSEIR.is_any_community_edge — Function

is_any_community_edge(g::SEIRGraph, e::AbstractEdge, c::Int)

Return true if at least one vertex of edge e is in community c (e is an edge related to community c).

GraphSEIR.is_within_community_edge — Function

is_within_community_edge(g::SEIRGraph, e::AbstractEdge, c::Int)

Return true if both vertices of edge e are in community c (e is an edge within community c.)

GraphSEIR.is_between_community_edge — Function

is_any_community_edge(g::SEIRGraph, e::AbstractEdge, c::Int)

Return true if only one vertex of edge e is in community c (e is an edge between community c and another community).

GraphSEIR.is_household_edge — Function

is_household_edge(g::SEIRGraph, e::AbstractEdge)

Return true if edge e is a household edge.

GraphSEIR.is_nonhousehold_edge — Function

is_nonhousehold_edge(g::SEIRGraph, e::AbstractEdge)

Return true if edge e is not a household edge.

GraphSEIR.is_any_community_nonhousehold_edge — Function

is_any_community_nonhousehold_edge(g::SEIRGraph, e::AbstractEdge, c::Int)

Return true if at least one vertex of edge e is in community c (e is an edge between community c and another community) and the edge is not a household edge.

Index

GraphSEIR.CommunityPairParameters

GraphSEIR.CommunityParameters

GraphSEIR.GlobalParameters

GraphSEIR.SEIRInfectionTree

GraphSEIR.SEIRParameters

GraphSEIR.SEIRTransitionHistory

GraphSEIR.CompartmentalTracker

GraphSEIR.R_i

GraphSEIR.add_new_infection!

GraphSEIR.community_eigenscores

GraphSEIR.community_expected_exposure

GraphSEIR.community_expected_exposures

GraphSEIR.community_expected_infection

GraphSEIR.community_expected_recovery

GraphSEIR.compute_R0

GraphSEIR.count

GraphSEIR.count

GraphSEIR.count

GraphSEIR.counts

GraphSEIR.counts_by_community

GraphSEIR.eigenscore

GraphSEIR.eigenvectors

GraphSEIR.epidemic_mobility_weight_matrix

GraphSEIR.expose!

GraphSEIR.expose_communities

GraphSEIR.expose_communities!

GraphSEIR.exposure_subgraph

GraphSEIR.format_data

GraphSEIR.full_transition!

GraphSEIR.get_betas

GraphSEIR.get_indices

GraphSEIR.get_indices

GraphSEIR.get_indices

GraphSEIR.initialize_with!

GraphSEIR.is_any_community_edge

GraphSEIR.is_any_community_nonhousehold_edge

GraphSEIR.is_between_community_edge

GraphSEIR.is_household_edge

GraphSEIR.is_nonhousehold_edge

GraphSEIR.is_within_community_edge

GraphSEIR.load_edgefile

GraphSEIR.load_graph

GraphSEIR.mdp_transition!

GraphSEIR.reset!

GraphSEIR.short_circuit_transition!

GraphSEIR.swap_edges!

GraphSEIR.transition!

GraphSEIR.transition_likelihood

GraphSEIR.vectorized_transition!

GraphSEIR.vertex_weight